Fix failed to publish asset for OpenRPC workflow
This commit partially reverts #3333. Specifically the workflow for publishing the OpenRPC document on a new release. The problem now is that with trigger `on.push.tags` there is a possibility that the actual release on Github isn't yet created.
Therefore, after generating the JSON file, uploading files to this non-existent release fails.

To mitigate the problem that #3333 tried to solve, I've changed the trigger back to `on.release.types` for the OpenRPC workflow. However, before the PR it was `[created]` and now I set this to `[published]`. This makes sure that even if the release initially was published as a draft, the workflow is still triggered.