PicoSync: Limits handling during request macro chain.
The request macro chain has a limit in terms of MAX_EPOCHS that can be returned.
From the pico sync point of view, this means that when we want to push the "latest" election block
we not necessary get the last one, but the latest one that we can get in a single request, so
pushing the latest election block is an iterative process.